Refactor async utils and use it for tasks too
